pyproject.toml
[project]
name = "metadata-filering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.14",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-neo4j>=0.9.0",
    "neo4j>=6.1.0",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.9",
]
